SplashScreenActivity
Criacao da acitivity pra splashscreen
