+++ app/src/main/java/com/gabrielcamargo/projetointegrador/MainActivity.kt
@@ -20,29 +20,6 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        // Fonte DK
-        val fontDK = ResourcesCompat.getFont(this, R.font.dk_butterfly_ball)
-        val textLogo: TextView = findViewById(R.id.txt_logo)
-        textLogo.typeface = fontDK
-
-        // ProgressBar
-        val maxSplashTime: Long = 3000
-        var progressSplash: Long = 0
-        val percentResult = maxSplashTime / 101
-        val progressSplashBar = findViewById<ProgressBar>(R.id.progressSplashBar)
-
-        for(x in 0..100){
-            Handler().postDelayed({
-                progressSplashBar.progress = x
-            }, progressSplash)
-            progressSplash = progressSplash + percentResult
-        }
-
-        Handler().postDelayed({
-            startActivity(Intent(this, LoginActivity::class.java))
-            finish()
-        },maxSplashTime)
-
     }
 
 }
