+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout  xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".SplashScreenActivity"
+    android:orientation="vertical"
+    android:weightSum="1"
+    android:padding="@dimen/dim_espacamento_40dp">
+
+    <ImageView
+        android:id="@+id/imageView4"
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        app:srcCompat="@mipmap/ic_channel_foreground"
+        android:layout_gravity="center"
+        android:layout_marginTop="@dimen/dim_espacamento_70dp"
+        android:layout_marginLeft="@dimen/dim_espacamento_15dp"/>
+
+    <TextView
+        android:id="@+id/txt_logo"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:text="@string/cinelist"
+        android:textSize="50sp"
+        android:textAlignment="center"
+        android:textColor="@color/white"
+        android:shadowRadius="1"
+        android:shadowDx="7"
+        android:shadowDy="7"
+        android:shadowColor="@color/gray_600"/>
+
+    <ProgressBar
+        android:id="@+id/progressSplashBar"
+        style="?android:attr/progressBarStyleHorizontal"
+        android:layout_width="match_parent"
+        android:layout_marginHorizontal="@dimen/dim_espacamento_40dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="bottom"
+        android:layout_weight="0.1" />
+
+</LinearLayout >
