+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,4 @@
-
-<resources>  
+<resources>
     <string name="app_name">Projeto Integrador</string>
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="image_film_list">Image - Film list</string>
@@ -10,10 +9,10 @@
     <string name="ten_stars_prefix">/ 10</string>
     <string name="nome_lista_exemplo">Minha lista de filmes</string>
     <string name="descricao_lista_exemplo">Exemplo de descrição da lista de filmes</string>
-   
-   <string name="filmes_populares">Filmes populares</string>
-   <string name="filmes_no_cinema">Filmes no cinema</string>
-  
+
+    <string name="filmes_populares">Filmes populares</string>
+    <string name="filmes_no_cinema">Filmes no cinema</string>
+
     <string name="sinopse">Sinopse</string>
     <string name="fotos">Fotos</string>
     <string name="reviews">Reviews</string>
